Refactor mission states to inherit generic functions
